+++ tests/test_watcher_schema.py
@@ -92,11 +92,11 @@ class WatcherSchemaTestCase(unittest.TestCase):
         assert len(config.escalation) == 2
         assert config.escalation[0] == dict(
             alarms=["Enabled.AT*"],
-            responders=[dict(name="stella", type="team")],
+            responder="stella",
             delay=0.11,
         )
         assert config.escalation[1] == dict(
             alarms=["Enabled.ATCamera"],
-            responders=[dict(name="someone@somewhere", type="user")],
+            responder="someone",
             delay=0.12,
         )